feat: fix send playernum
