+++ frontend/src/components/RandomWordButton.js
@@ -2,6 +2,7 @@ import axios from 'axios';
 import React, { useEffect } from 'react';
 
 const baseURL = 'http://localhost:5000/randwords';
+// const baseURL = 'http://127.0.0.1:5000/randwords';
 function RandomWordButton({ image, setWord }) {
   async function onClick() {
     const response = await axios.get(baseURL);
