+++ frontend/src/pages/Main.js
@@ -2,7 +2,6 @@ import React from 'react';
 import MainText from '../components/MainText';
 import AboutIcon from '../assets/icons/aboutIcon.png';
 import MainIcon from '../assets/icons/mainIcon.png';
-import SetNum from '../components/SetNum';
 import GameStartButton from '../components/GameStartButton';
 
 /*
@@ -19,21 +18,17 @@ function Main() {
 
         <img className=" absolute mt-[13vh] right-10  h-[35rem]  " src={MainIcon} alt="" />
 
-        <div className=" flex-col ml-[13vh] mt-[28vh] flex ">
+        <div className=" flex-col ml-[13vh] mt-[30vh] flex ">
           <MainText text=" 두들, 두들! " />
           <h2 className="className = mt-8 ml-52 text-5xl text-primary-1 font-cookierun doddleshadow textborder">
             Doodle, Doodle
           </h2>
 
-          <div className="flex-row flex left-0 mt-[16vh] ml-[5.5vw]">
+          <div className="flex-row flex left-0 mt-[14vh] ml-[6vw]">
             <div className="flex ">
               <GameStartButton />
-              <div className="ml-20">
-                <SetNum />
-              </div>
             </div>
           </div>
-          
         </div>
       </div>
     </div>
